+++ hugegraph-mcp/hugegraph_mcp/server.py
@@ -378,6 +378,8 @@ def import_graph_data_tool(
     dry_run: bool = True,
     confirm: bool = False,
     plan_hash: str | None = None,
+    nonce: str | None = None,
+    expires_at: float | None = None,
 ) -> dict:
     """兼容图数据导入入口 — 推荐使用 extract_graph_data_tool。
 
@@ -403,6 +405,8 @@ def import_graph_data_tool(
             dry_run=dry_run,
             confirm=confirm,
             plan_hash=plan_hash,
+            nonce=nonce,
+            expires_at=expires_at,
         )
 
     if mode == "table":
